// ui/src/app/model/JobAbortResponse.ts
/**
 * Job Monitor Service
 * Job Monitor API for interacting with asynchronous batch jobs and workflows.
 *
 * OpenAPI spec version: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

import * as models from './models';

export interface JobAbortResponse {
    /**
     * The identifier of the job
     */
    id: string;

    /**
     * The status of the job
     */
    status: string;

}